Refactor Code at root/tabs.py sent to root/views.py + deleted root/tabs.py
