+++ tkinter_mvc_app/root/views.py
@@ -1,6 +1,7 @@
 import tkinter as tk
+from tkinter import ttk
 
-from tkinter_mvc_app.root.tabs import Tab
+from tkinter_mvc_app.books.views.frames import BookFrame
 from tkinter_mvc_app.helpers.const import APP_NAME, WIDTH, HEIGHT
 
 
@@ -16,12 +17,27 @@ class GUISetup(tk.Tk):
         self.minsize(min_width, min_height)
 
 
+class Tabs(ttk.Notebook):
+    def __init__(self, parent):
+        super().__init__(parent)
+
+        # Create book tab
+        self.book_tab = BookFrame(self)
+        self.book_tab.grid(row=0, column=0, sticky="nsew")
+        self.add(self.book_tab, text="Libros")
+
+        # Create another tab
+        self.another_tab = BookFrame(self)
+        self.another_tab.grid(row=0, column=0, sticky="nsew")
+        self.add(self.another_tab, text="Otra pestaña")
+
+
 class MainView:
     def __init__(self):
         self.gui_setup = GUISetup(APP_NAME, WIDTH, HEIGHT)
 
         # Tabs
-        self.gui_setup.tabs = Tab(self.gui_setup)
+        self.gui_setup.tabs = Tabs(self.gui_setup)
         self.gui_setup.tabs.pack(side="top", fill="x")
 
     def run_mainloop(self):
